Updated roborio library to support software-level yaw resets, matching the C++ code.
Also updated the C++ code to be as identical as possible to the Java
code.

+++ roborio/c++/navx_frc_cpp/include/AHRS.h
@@ -131,7 +131,7 @@ private:
 	
 	bool enable_boardlevel_yawreset;
     bool previously_connected;
-    long last_yawreset_request_timestamp;
+    double last_yawreset_request_timestamp;
     uint32_t successive_suppressed_yawreset_request_count;
 
 public:
